+++ lib/jazzy/gem_version.rb
@@ -1,3 +1,3 @@
 module Jazzy
-  VERSION = '0.2.4' unless defined? Jazzy::VERSION
+  VERSION = '0.3.2' unless defined? Jazzy::VERSION
 end